Added the test suite to new configuration
